requirements.txt
numpy
scipy
pywavelets
soundfile
cryptography